// src/pages/CienciaTecnologiaInnovacion/types.ts
export interface Fundamento {
  title: string;
  text: string;
}

export interface TipoItem {
  label: string;
  desc: string;
}

export interface ImpactoItem {
  icon: string;
  title: string;
  text: string;
}

export interface Politica {
  num: string;
  title: string;
  text: string;
}

export interface Tendencia {
  title: string;
  text: string;
  color: string;
}